Fixed ListLayout Issue In Pagination
Hey Tim, I've encountered a small bug which was passing the un-sorted blogs to the ListLayout on the `./blog/[page].tsx` page. I've added the `sortedBlogPost()` to `getStaticProps()` so that the sorted blog posts are passed to the ListLayout Component. 

It's my first time contributing, do let me know if this requires any further clarification. Thanks for all your hard work in bringing us this template!